+++ src/utils/index.ts
@@ -7,3 +7,4 @@
  */
 export * from "./format";
 export * from "./event";
+export * from "./game";